+++ packages/sdk-bridge/src/BridgeMessage.ts
@@ -3,7 +3,7 @@ import { arrayify, hexlify } from '@ethersproject/bytes';
 import { TransactionReceipt } from '@ethersproject/abstract-provider';
 import { ethers } from 'ethers';
 import * as bridge from '@nomad-xyz/contracts-bridge';
-import { NomadMessage, AnnotatedDispatch } from '@nomad-xyz/sdk';
+import { NomadMessage, Dispatch } from '@nomad-xyz/sdk';
 import { ResolvedTokenInfo, TokenIdentifier } from './tokens';
 import { BridgeContracts } from './BridgeContracts';
 import { BridgeContext } from './BridgeContext';
@@ -82,14 +82,14 @@ export class BridgeMessage extends NomadMessage<BridgeContext> {
    */
   constructor(
     context: BridgeContext,
-    event: AnnotatedDispatch,
+    dispatch: Dispatch,
     token: TokenIdentifier,
     callerKnowsWhatTheyAreDoing: boolean,
   ) {
     if (!callerKnowsWhatTheyAreDoing) {
       throw new Error('Use `fromReceipt` to instantiate');
     }
-    super(context, event);
+    super(context, dispatch);
 
     const fromBridge = context.mustGetBridge(this.message.from);
     const toBridge = context.mustGetBridge(this.message.destination);
@@ -124,18 +124,16 @@ export class BridgeMessage extends NomadMessage<BridgeContext> {
    * Attempt to instantiate some BridgeMessages from a transaction receipt
    *
    * @param context The {@link NomadContext} to use.
-   * @param nameOrDomain the domain on which the receipt was logged
    * @param receipt The receipt
    * @returns an array of {@link BridgeMessage} objects
    * @throws if any message cannot be parsed as a bridge message
    */
-  static fromReceipt(
+  static async fromReceipt(
     context: BridgeContext,
-    nameOrDomain: string | number,
     receipt: TransactionReceipt,
-  ): AnyBridgeMessage[] {
+  ): Promise<AnyBridgeMessage[]> {
     const nomadMessages: NomadMessage<BridgeContext>[] =
-      NomadMessage.baseFromReceipt(context, nameOrDomain, receipt);
+      await NomadMessage.baseFromReceipt(context, receipt);
     const bridgeMessages: AnyBridgeMessage[] = [];
     for (const nomadMessage of nomadMessages) {
       try {
@@ -155,20 +153,17 @@ export class BridgeMessage extends NomadMessage<BridgeContext> {
    * Attempt to instantiate EXACTLY one BridgeMessage from a transaction receipt
    *
    * @param context The {@link BridgeContext} to use.
-   * @param nameOrDomain the domain on which the receipt was logged
    * @param receipt The receipt
    * @returns an array of {@link BridgeMessage} objects
    * @throws if any message cannot be parsed as a bridge message, or if there
    *         is not EXACTLY 1 BridgeMessage in the receipt
    */
-  static singleFromReceipt(
+  static async singleFromReceipt(
     context: BridgeContext,
-    nameOrDomain: string | number,
     receipt: TransactionReceipt,
-  ): AnyBridgeMessage {
-    const messages: AnyBridgeMessage[] = BridgeMessage.fromReceipt(
+  ): Promise<AnyBridgeMessage> {
+    const messages: AnyBridgeMessage[] = await BridgeMessage.fromReceipt(
       context,
-      nameOrDomain,
       receipt,
     );
     if (messages.length !== 1) {
@@ -183,7 +178,7 @@ export class BridgeMessage extends NomadMessage<BridgeContext> {
    *
    * @param context The {@link NomadContext} to use.
    * @param nameOrDomain the domain on which the receipt was logged
-   * @param transactionHash The transaction hash
+   * @param transactionHash the transaction hash on the origin chain
    * @returns an array of {@link BridgeMessage} objects
    * @throws if any message cannot be parsed as a bridge message
    */
@@ -197,7 +192,7 @@ export class BridgeMessage extends NomadMessage<BridgeContext> {
     if (!receipt) {
       throw new Error(`No receipt for ${transactionHash} on ${nameOrDomain}`);
     }
-    return BridgeMessage.fromReceipt(context, nameOrDomain, receipt);
+    return BridgeMessage.fromReceipt(context, receipt);
   }
 
   /**
@@ -206,7 +201,7 @@ export class BridgeMessage extends NomadMessage<BridgeContext> {
    *
    * @param context The {@link NomadContext} to use.
    * @param nameOrDomain the domain on which the receipt was logged
-   * @param transactionHash The transaction hash
+   * @param transactionHash the transaction hash on the origin chain
    * @returns an array of {@link BridgeMessage} objects
    * @throws if any message cannot be parsed as a bridge message, or if there is
    *         not EXACTLY one such message
@@ -221,7 +216,7 @@ export class BridgeMessage extends NomadMessage<BridgeContext> {
     if (!receipt) {
       throw new Error(`No receipt for ${transactionHash} on ${nameOrDomain}`);
     }
-    return BridgeMessage.singleFromReceipt(context, nameOrDomain, receipt);
+    return BridgeMessage.singleFromReceipt(context, receipt);
   }
 
   /**
@@ -272,10 +267,10 @@ export class TransferMessage extends BridgeMessage {
 
   constructor(
     context: BridgeContext,
-    event: AnnotatedDispatch,
+    dispatch: Dispatch,
     parsed: ParsedTransferMessage,
   ) {
-    super(context, event, parsed.token, true);
+    super(context, dispatch, parsed.token, true);
     this.action = parsed.action;
   }