+++ packages/sdk-bridge/src/index.ts
@@ -1,8 +1,5 @@
 export { BridgeContracts } from './BridgeContracts';
 
-export * from './bridgeEvents';
-export * as bridgeEvents from './bridgeEvents';
-
 export * from './BridgeContext';
 
 export * from './BridgeMessage';