+++ packages/sdk/src/index.ts
@@ -2,9 +2,8 @@ export { CoreContracts } from './CoreContracts';
 
 export * from './messages/NomadMessage';
 
-export type { AnnotatedLifecycleEvent, NomadLifecyleEvent } from './events';
-export * from './events';
-
 export { FailedHomeError } from './error';
 
 export { NomadContext } from './NomadContext';
+
+export * from './messages/types';