+++ packages/sdk/src/messages/types.ts
@@ -1,0 +1,57 @@
+import { BigNumber } from '@ethersproject/bignumber';
+import { TransactionReceipt } from '@ethersproject/abstract-provider';
+
+// copied from Home.d.ts
+export type Dispatch = {
+  args: {
+    messageHash: string;
+    leafIndex: BigNumber;
+    destinationAndNonce: BigNumber;
+    committedRoot: string;
+    message: string;
+  };
+  transactionHash: string;
+  receipt: TransactionReceipt;
+};
+
+export type ParsedMessage = {
+  from: number;
+  sender: string;
+  nonce: number;
+  destination: number;
+  recipient: string;
+  body: string;
+};
+
+export enum MessageStatus {
+  dispatched = 0,
+  included,
+  relayed,
+  updated,
+  received,
+  processed,
+}
+
+export enum ReplicaStatusNames {
+  None = 'none',
+  Proven = 'proven',
+  Processed = 'processed',
+}
+
+type ReplicaMessageStatusNone = {
+  status: ReplicaStatusNames.None;
+};
+
+type ReplicaMessageStatusProcess = {
+  status: ReplicaStatusNames.Processed;
+};
+
+type ReplicaMessageStatusProven = {
+  status: ReplicaStatusNames.Proven;
+  root: string;
+};
+
+export type ReplicaMessageStatus =
+  | ReplicaMessageStatusNone
+  | ReplicaMessageStatusProcess
+  | ReplicaMessageStatusProven;